worker/requirements.txt
kafka-python
sqlalchemy
asyncpg
pgvector
pdfplumber
python-docx
fugashi[unidic-lite]
langdetect
voyageai
cohere
python-dotenv
pydantic-settings
